[#241] Clamped frame dimensions to the GIF 16-bit limit and shared the test GIF parser.
